requirements.txt
fastapi[standard]
uvicorn
google-generativeai
pydantic
python-dotenv